// frontend/src/client/types.gen.ts
export type LossCodeCategoryInHierarchy = {
    name: string;
    display_order?: number;
    id: string;
    type_id: string;
    loss_codes?: Array<LossCodeInHierarchy>;
};
// ...
export type LossCodeInHierarchy = {
    id: string;
    code: string;
    name: string;
};

export type LossCodePublic = {
    id: string;
    code: string;
    name: string;
    category_id: string;
};

export type LossCodesPublic = {
    data: Array<LossCodePublic>;
    count: number;
};

export type LossCodeTypePublic = {
    key: string;
    label: string;
    color?: string;
    display_order?: number;
    id: string;
};

export type LossCodeTypeWithHierarchy = {
    key: string;
    label: string;
    color?: string;
    display_order?: number;
    id: string;
    categories?: Array<LossCodeCategoryInHierarchy>;
};